Update to use seperate modules for different  schemas
